+++ src/plugin.ts
@@ -213,6 +213,7 @@ export function vueInternationalization(options?: Partial<VueInternationalizatio
 						global: globalMessages[currentOptions.primaryLocale],
 						messageSyntax: currentOptions.messageSyntax,
 						transformAll: currentOptions.sfcTransform === 'all',
+						moduleExpression: JSON.stringify(toRuntimeModuleId(id, root)),
 					});
 
 			if (!transformed) {